fix illegal instruction in std::ios_base::Init::Init() by disabling llvm static linking with libstdc++
I bisected this back to https://github.com/rust-lang/rust/pull/94832 which made static linking with libstdc++ the default. Unfortunately this means we link with the host copy, which is not i586-compiled, so it crashes on the target machine. Instead, dynamically link to the target machine's, which will be compiled properly.